BIG updated for digistores. Expanded for future multi-item digistores and digistore chests. Added search and ordering to digistore manager, as well as some performance improvments with fluid tanks. Still working on digistore insert/extract logic.

+++ src/main/java/theking530/common/gui/drawables/IDrawable.java
@@ -1,5 +1,5 @@
 package theking530.common.gui.drawables;
 
 public interface IDrawable {
-	public void draw(int x, int y);
+	public void draw(float x, float y);
 }
